12.secret_share/requirements.txt
pycrypto==2.6.1
numpy==1.15.4